feat: create loader
